+++ src/elements/actions/SetStatus.php
@@ -72,7 +72,7 @@ class SetStatus extends ElementAction
     /**
      * @inheritdoc
      */
-    public function getTriggerHtml()
+    public function getTriggerHtml(): ?string
     {
         return Craft::$app->getView()->renderTemplate('_components/elementactions/SetStatus/trigger');
     }