+++ src/elements/actions/View.php
@@ -45,7 +45,7 @@ class View extends ElementAction
     /**
      * @inheritdoc
      */
-    public function getTriggerHtml()
+    public function getTriggerHtml(): ?string
     {
         $type = Json::encode(static::class);